refactor: move x/container out of internal/ (#336)
Pure import-path rewrite — first commit of PRD #335. Relocates
internal/x/container/ to x/container/ and updates all importers.
No type, signature, or behaviour changes.

- internal/x/container/{container,singleton,closer,hz}.go and tests
  moved verbatim to x/container/.
- All imports of github.com/kannon-email/kannon/internal/x/container
  rewritten to github.com/kannon-email/kannon/x/container across cmd/,
  pkg/, and e2e/.
- ARCHITECTURE.md path reference updated.

go build, go vet, make test, and make lint all green.
