+++ pkg/hz/hz.go
@@ -3,7 +3,7 @@ package hz
 import (
 	"context"
 
-	"github.com/kannon-email/kannon/internal/x/container"
+	"github.com/kannon-email/kannon/x/container"
 )
 
 type HZService interface {
