+++ x/container/closer_test.go
