+++ x/container/singleton.go
