+++ x/container/singleton_test.go
